Remove code duplication, fix process endtime tests and logic
Update coverage requirements
