+++ package.json
@@ -63,10 +63,10 @@
     ],
     "coverageThreshold": {
       "global": {
-        "branches": 0,
-        "functions": 1,
-        "lines": 1,
-        "statements": 1
+        "branches": 25,
+        "functions": 13,
+        "lines": 25,
+        "statements": 25
       }
     }
   },
